Bug 635007: puppet support for installing virtualenvs full of packages
This installs python packages by using pip to read from the puppet http root -
packages are not stored on the slaves, and only the packages required are
downloaded.  It uses pip-check.py to check for a package in the virutalenv, and
only tries to install if that package is not found -- so no action is taken in
the normal (already-installed) case.

r=bhearsum